requirements.txt
numpy==1.26.4
python-doctr==0.11.0
torch==2.4.1
nltk==3.9.1
python-dotenv==1.0.1
opencv-python==4.10.0.84
transformers>=4.40.0
sentence-transformers>=2.6.0
langchain==0.3.3
langchain-community==0.3.2
langchain-text-splitters==0.3.0
chromadb==0.5.15
PyMuPDF>=1.24.0
camelot-py[cv]>=0.11.0
pandas==2.2.3
pytesseract>=0.3.10
pillow==10.4.0
fastapi==0.115.2
uvicorn>=0.27.0
reportlab>=4.1.0
pytesseract==0.3.10
packaging==24.1
accelerate